moved connectionSettings to connection.go
